Reccomend completing curl before exec
There's a vague argument that it's very slightly safer (both in the "unexpected error" sense [and the "security" sense](https://unix.stackexchange.com/a/339276)) to ensure that `curl` has fully downloaded a shell-script before invoking `sfh`.

I've updated the README according to this guidance, but I didn't check if there's any other documentation with the `curl | sh` pattern; feel free to update my branch if you know of any. (=

It's a tiny tweak, and the benefits are arguable, so I certainly won't be offended if you don't want to merge. Security theatre benefits nobody. `¯\_(ツ)_/¯`